+++ client/bitmap.h
@@ -4,69 +4,72 @@
 #include <QImage>
 #include <QPainter>
 #include <QString>
+#include <QThread>
 #include <map>
-#include "../common/event.h"
+#include <common/event.h>
+#include <common/console.h>
+#include <common/eventdispatcher.h>
+#include <common/matrix.h>
 #include "interactiveobject.h"
-#include "QThread"
-#include "../common/console.h"
 
-class TextureLoader : public QObject, public EventDispatcher
-{
-    Q_OBJECT
+class TextureLoader : public QObject, public EventDispatcher {
+Q_OBJECT
 public:
-    TextureLoader(QString path)
-    {
+    TextureLoader(QString path) {
         this->path = path;
         thread = new QThread();
         QObject::connect(thread, SIGNAL(started()), this, SLOT(run()));
         texture = 0;
     }
 
-    ~TextureLoader()
-    {
+    ~TextureLoader() {
         thread->deleteLater();
     }
 
-    void start()
-    {
+    void start() {
         this->moveToThread(thread);
         thread->start();
     }
 
-    QImage* getTexture()
-    {
+    QImage *getTexture() {
         return texture;
     }
 
 public slots:
-    void run()
-    {
-        QImage* temp = new QImage();
+
+    void run() {
+        QImage *temp = new QImage();
         temp->load(this->path);
         thread->quit();
-        texture =  temp;
+        texture = temp;
         dispatchEvent(Event(this, Event::COMPLETE));
     }
+
 private:
-    QImage* texture;
+    QImage *texture;
     QString path;
-    QThread* thread;
+    QThread *thread;
 };
 
-class Bitmap : public InteractiveObject, public EventHandler
-{
+class Bitmap : public InteractiveObject, public EventHandler {
 public:
     Bitmap();
+
     ~Bitmap();
-    QImage* getSource() const;
+
+    QImage *getSource() const;
+
     void load(QString);
-    void render(QPainter*, const Matrix&, bool, float);
+
+    void render(QPainter *, const Matrix &, bool, double);
+
     void Invoke(const Event &event);
+
 private:
-    static std::map<QString, QImage*> cache;
-    QImage* source;
+    static std::map<QString, QImage *> cache;
+    QImage *source;
     QString source_path;
-    TextureLoader* loader;
+    TextureLoader *loader;
 };
 
 #endif // BITMAP_H
