+++ multicastparams.cpp
@@ -1,17 +1,14 @@
 #include "multicastparams.h"
 
-MulticastParams::MulticastParams()
-{
+MulticastParams::MulticastParams() {
 
 }
 
-QHostAddress MulticastParams::getGroupAddress()
-{
+QHostAddress MulticastParams::getGroupAddress() {
     return QHostAddress("239.255.43.21");
 }
 
-int MulticastParams::getGroupPort()
-{
+quint16 MulticastParams::getGroupPort() {
     return 45454;
 }
